Add Basic Template to modularize game files.
Added project and file structure to create a basic template and
encapsulate the respective code. The code is divided in two different
folders; scenes and config. The config folder as the suggests, contains
the configuration for Phaser. The scenes folder contains different
scenes for the game. Namely, boot, preload, credits, game-scenes,
option-scenes, title.

ES6 export and import is used. Babel is used to transpile.